🧹 Refactor TagAutocompleteInput component into smaller parts
